feat: Add test for scheduling client execution action
The commit adds a new test `test_schedule_action` in the `client_execution_action.rs` file. This test is responsible for scheduling a client execution action using the `ControllerClient` and asserting the response action ID. This test ensures the proper functioning of the scheduling feature and improves the overall code quality.

Signed-off-by: benoit-planche <benoit.planche@etu.umontpellier.fr>