refactor: Add Action struct and ActionQueue implementation
This commit adds the `Action` struct and implements the `ActionQueue` collection in the `controller_logic.rs` file. The `Action` struct represents an action in the queue and contains an ID, a score, and additional fields from the `ActionRequest` proto. The `ActionQueue` is a collection of `Actions` stored in a vector.

Signed-off-by: Théo Tchilinguirian <theo.tchlx@gmail.com>

Co-authored-by: benoit-planche <benoit.planche@etu.umontpellier.fr>