refactor: Rename Pipeline struct to PipelineYaml
This commit renames the `Pipeline` struct to `PipelineYaml` to better reflect its purpose of representing a YAML configuration for pipelines. This change is consistent with the recent updates to the `ManifestParser` trait and the `MockManifestParser` implementation.

Note: This commit message follows the established convention of using the "refactor" type for code restructuring or renaming.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>