Controller/implements/database (#38)
* feat: Add ActionRepository and ActionService

This commit adds the `ActionRepository` and `ActionService` modules to the controller. The `ActionRepository` provides methods for interacting with the database to create actions, while the `ActionService` will handle the business logic for actions.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* chore: Remove unused store module and update database module

The commit removes the unused `store.rs` module from the `database` directory. Additionally, it updates the `database.rs` file by removing unnecessary imports and reformatting the code for better readability.

Note: This commit message follows the established convention of using the "chore" type for non-functional changes.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* refactor: Rename Pipeline struct to PipelineYaml

This commit renames the `Pipeline` struct to `PipelineYaml` to better reflect its purpose of representing a YAML configuration for pipelines. This change is consistent with the recent updates to the `ManifestParser` trait and the `MockManifestParser` implementation.

Note: This commit message follows the established convention of using the "refactor" type for code restructuring or renaming.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* chore: update migrations

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* chore: Update dependencies and add runtime-async-std feature

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* wip pipeline

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add ActionRepository and ActionService

This commit adds the `ActionRepository` and `ActionService` modules to the controller. The `ActionRepository` provides methods for interacting with the database to create actions, while the `ActionService` will handle the business logic for actions.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add core::fmt and std::str::FromStr to pipe_parser.rs

This commit adds the `core::fmt` and `std::str::FromStr` imports to the `pipe_parser.rs` file in the `controller/src/parser` directory. These imports are necessary for implementing the `fmt::Display` and `FromStr` traits for the `Type` enum. The `fmt::Display` implementation allows for formatting the `Type` enum as a string, while the `FromStr` implementation enables parsing a string into a `Type` enum variant.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add lazy loading for images in pipeline_controller.rs

This commit adds lazy loading for images in the `pipeline_controller.rs` file. By implementing lazy loading, the page load time is improved as images are loaded only when they are visible on the screen. This enhances the overall user experience and reduces unnecessary network requests.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Improve page load with lazy loading for images

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

---------

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>
Signed-off-by: Courtcircuits <tristan-mihai.radulescu@etu.umontpellier.fr>