feat: Add find_by_id and find_by_pipeline_id methods to ActionRepository
The commit adds the `find_by_id` and `find_by_pipeline_id` methods to the `ActionRepository` in the `action_repository.rs` file. These methods allow for retrieving actions from the database based on their ID or pipeline ID. This functionality enhances the flexibility and querying capabilities of the repository.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>