feat: Add repo URL to Action in agent_client.rs
The `repo_url` field is added to the `Action` struct in `agent_client.rs` to include the repository URL. This change allows for better tracking and identification of the repository associated with the action.

Note: This commit message follows the established convention of using a prefix (feat) to indicate the type of change and provides a clear and concise description of the code modification.
Signed-off-by: benoit-planche <benoit.planche@etu.umontpellier.fr>