feat: Add scheduler client module
This commit adds a new module called `scheduler_client` to the `domain/services` directory. The `scheduler_client` module contains the `SchedulerClient` trait, which provides a method for scheduling actions. This module enables communication with the scheduler service and enhances the functionality of the controller.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>