feat: Add services module to domain
The commit adds a new `services` module to the `domain` module in the controller. This module includes the `scheduler_client` module and the `SchedulerClient` trait, which provides a method for scheduling actions. The `services` module enhances the functionality of the controller by enabling communication with the scheduler service.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>