Create domain architecture hexa (#56)
* feat: Add domain module and entities for actions, commands, logs, and pipelines

This commit adds a new `domain` module to the controller, which includes sub-modules for `entities`. The `entities` module contains the following files: `action.rs`, `command.rs`, `log.rs`, `pipeline.rs`, and `scheduler.rs`. These files define the data structures and error types related to actions, commands, logs, and pipelines. This modular structure improves code organization and enhances the maintainability of the controller.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add repositories module to domain

This commit adds a new `repositories` module to the `domain` module in the controller. The `repositories` module includes sub-modules for `action_repository`, `command_repository`, `log_repository`, and `pipeline_repository`. These repositories provide methods for interacting with the database to perform CRUD operations on actions, commands, logs, and pipelines. This modular structure enhances code organization and improves the maintainability of the controller.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add scheduler client module

This commit adds a new module called `scheduler_client` to the `domain/services` directory. The `scheduler_client` module contains the `SchedulerClient` trait, which provides a method for scheduling actions. This module enables communication with the scheduler service and enhances the functionality of the controller.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add services module to domain

The commit adds a new `services` module to the `domain` module in the controller. This module includes the `scheduler_client` module and the `SchedulerClient` trait, which provides a method for scheduling actions. The `services` module enhances the functionality of the controller by enabling communication with the scheduler service.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

---------

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>
Signed-off-by: Courtcircuits <tristan-mihai.radulescu@etu.umontpellier.fr>