feat: Add GrpcSchedulerClient for interacting with the scheduler service
This commit adds the `GrpcSchedulerClient` struct to the `grpc_scheduler_client.rs` file in the `controller` project. The `GrpcSchedulerClient` implements the `SchedulerClient` trait and provides methods for scheduling actions with the scheduler service. This client enhances the functionality of the controller by enabling communication with the scheduler service.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>