feat: Add app_context module and related structs
This commit adds the `app_context.rs` file to the `application` module in the `controller` project. The `app_context.rs` file defines the `AppContext` struct, which represents the application context. This struct will be used to store shared data and services throughout the application. The addition of the `AppContext` module enhances the modularity and organization of the codebase.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>