feat: Add command_service module and implementation
This commit adds the `command_service.rs` file to the `services` module in the `application` directory. The `command_service.rs` file contains the implementation of the `CommandService` trait, which provides methods for finding commands by action ID and creating new commands. This module enhances the functionality of the controller by enabling interaction with commands in the application.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>