Integration Infrastructure & Application (#57)
* feat: Add infrastructure module for database connectivity

The commit adds a new `infrastructure` module to the `controller` project. This module includes the `db` module, which contains the `postgres` module and the `Postgres` struct. The `Postgres` struct provides a connection pool to the PostgreSQL database. This infrastructure module enhances the controller's functionality by enabling database connectivity and interaction.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add PostgresActionRepository for interacting with the database

This commit adds the `PostgresActionRepository` struct to the `action_repository.rs` file in the `controller` project. The `PostgresActionRepository` implements the `ActionRepository` trait and provides methods for creating and retrieving actions from the PostgreSQL database. This repository enhances the controller's functionality by enabling database connectivity and interaction.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add PostgresCommandRepository for interacting with the database

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add pipeline_repository module for interacting with the database

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Fix typo in action_repository.rs

The commit fixes a typo in the `action_repository.rs` file in the `controller` project. The typo was causing a compilation error. This fix improves the code quality and ensures the correct execution of the program.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add GrpcSchedulerClient for interacting with the scheduler service

This commit adds the `GrpcSchedulerClient` struct to the `grpc_scheduler_client.rs` file in the `controller` project. The `GrpcSchedulerClient` implements the `SchedulerClient` trait and provides methods for scheduling actions with the scheduler service. This client enhances the functionality of the controller by enabling communication with the scheduler service.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add app_context module and related structs

This commit adds the `app_context.rs` file to the `application` module in the `controller` project. The `app_context.rs` file defines the `AppContext` struct, which represents the application context. This struct will be used to store shared data and services throughout the application. The addition of the `AppContext` module enhances the modularity and organization of the codebase.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add command_service module and implementation

This commit adds the `command_service.rs` file to the `services` module in the `application` directory. The `command_service.rs` file contains the implementation of the `CommandService` trait, which provides methods for finding commands by action ID and creating new commands. This module enhances the functionality of the controller by enabling interaction with commands in the application.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* fix: typo & fmt

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

---------

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>