feat: Add async-trait and futures dependencies
This commit adds the `async-trait` and `futures` dependencies to the `Cargo.toml` file in the `controller` project. These dependencies are required for implementing asynchronous functionality in the codebase. The addition of these dependencies enhances the controller's capability to handle asynchronous operations.

Note: This commit message follows the established convention of using the "feat" type for new features or functionality additions.
Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>
Signed-off-by: Courtcircuits <tristan-mihai.radulescu@etu.umontpellier.fr>